create single main function
